First hack at a Finder service to add selection to IPFS.
Currently using a bash script to call ipfs, a javascript to display the
resulting hashes in a window and AppleScript to copy the selected hash
to the system clipboard and Automator to send all the hashes and file
names to the notifications panel/popover.

Does not handle directory selections very well. It adds the contents
just fine but displays all the content's hashes on one line.
I'll fix that by using something other than Bash for handling the
output.